compact tables as a postprocessing step; several fixes and improvements. bench: 5752714, hash: B5794B30EDB27553

+++ motors/src/search/multithreading.rs
@@ -212,7 +212,7 @@ impl<B: Board> AtomicSearchState<B> {
     }
 
     pub fn iterations(&self) -> DepthPly {
-        DepthPly::new(self.iteration.load(Relaxed) as usize)
+        DepthPly::new(self.iteration.load(Relaxed))
     }
 
     pub fn seldepth(&self) -> DepthPly {
